requirements.txt
appdirs==1.4.4
astroid==2.4.1
autopep8==1.5.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
IMAPClient==2.1.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
openpyxl==2.1.4
pbr==5.4.5
Pillow==7.1.2
PyAutoGUI==0.9.50
pycodestyle==2.6.0
PyGetWindow==0.0.8
pytest==5.4.3
PyJWT==1.7.1
pylint==2.5.2
PyMsgBox==1.0.8
PyPDF2==1.26.0
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-docx==0.8.10
PyTweening==1.0.3
pytz==2020.1
pyzmail36==1.0.4
requests==2.23.0
rubicon-objc==0.3.1
schedule==0.6.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
stevedore==1.32.0
toml==0.10.1
twilio==6.40.0
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
python-decouple==3.1
tzwhere==3.0.3
python-dateutil==2.8.1
